windows: Absolute references to C:\ should not prepend Files/
Absolute symlinks for binares like C:\bin\busybox.exe do not work
properly if the link name gets prepended with /Files. It will show up
as:

date.exe -> [Files\C:\bin\busybox.exe]

When Files/ is not prepended the resulting links work correctly when
the container is deployed to server.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>